[rename] supabase.js => ComunitySupabase.js 변경
