+++ src/api/supabase/supabase.js
@@ -1,5 +1,5 @@
-// import { createClient } from '@supabase/supabase-js';
+import { createClient } from '@supabase/supabase-js';
 
-// const supabaseUrl = 'https://rtjzvtuqyafegkvoirwc.supabase.co';
-// const supabaseKey = process.env.REACT_APP_SUPABASE_KEY;
-// export const supabase = createClient(supabaseUrl, supabaseKey);
+const supabaseUrl = 'https://rtjzvtuqyafegkvoirwc.supabase.co';
+const supabaseKey = process.env.REACT_APP_SUPABASE_KEY;
+export const supabase = createClient(supabaseUrl, supabaseKey);
